Add gallery page with ArtworksList in it
