+++ sonar-project.properties
@@ -9,5 +9,5 @@ sonar.sources=./src
 sonar.sourceEncoding=UTF-8
 
 sonar.test.inclusions=src/**/*.test.ts,src/**/*.test.tsx
-sonar.coverage.exclusions=**/src/main.tsx,**/src/redux/hooks.ts,**/src/redux/store.ts,**/src/setupTest.ts, **/src/components/App/App.tsx
+sonar.coverage.exclusions=**/src/main.tsx,**/src/redux/hooks.ts,**/src/redux/store.ts,**/src/setupTest.ts, **/src/components/App/App.tsx, **/src/pages/*.tsx
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
